updates to library structure
